+++ flow_app/lib/screen/AppConfig/change_successful.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../main/app_config_screen.dart';
+import '../main/account/app_config_screen.dart';
 
 class ChangeSuccessfulScreen extends StatelessWidget {
   const ChangeSuccessfulScreen({
