+++ flow_app/lib/screen/apts/contact_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../edit_apt/create_schedule_screen.dart';
+import 'edit_apt/create_schedule_screen.dart';
 
 class ContactScreen extends StatefulWidget {
   const ContactScreen(
