+++ flow_app/lib/screen/apts/edit_apt/create_log_screen.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:snackbar/snackbar.dart';
 
-import '../../components/Inputs/date_picker_rectangle.dart';
-import '../../components/Inputs/limited_text_input_field.dart';
-import '../../components/Inputs/name_input_field.dart';
-import '../../components/Inputs/price_input_field.dart';
-import '../../components/Inputs/time_picker_rectangle.dart';
+import '../../../components/Inputs/date_picker_rectangle.dart';
+import '../../../components/Inputs/limited_text_input_field.dart';
+import '../../../components/Inputs/name_input_field.dart';
+import '../../../components/Inputs/price_input_field.dart';
+import '../../../components/Inputs/time_picker_rectangle.dart';
 
-//TODO: check if client exists
+//TODO: http request to check if client exists
 class CreateLogScreen extends StatefulWidget {
   const CreateLogScreen({
     super.key,
