+++ flow_app/lib/screen/main/account/app_config_screen.dart
@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
 
-import '../../components/Buttons/colored_border_text_button.dart';
-import '../../components/Buttons/custom_button.dart';
-import '../../components/Inputs/check_input_field.dart';
-import '../../components/Inputs/enum_field.dart';
-import '../../components/Inputs/time_picker_rectangle.dart';
-import '../../components/warning_modal.dart';
-import '../AppConfig/change_password/change_password_screen.dart';
-import '../AppConfig/change_phone/change_phone_screen.dart';
-import '../AppConfig/deactivate_account/deactivate_account_screen.dart';
-import '../AppConfig/delete_account/delete_warning_screen.dart';
-import '../auth/initial_screen.dart';
+import '../../../components/Buttons/colored_border_text_button.dart';
+import '../../../components/Buttons/custom_button.dart';
+import '../../../components/Inputs/check_input_field.dart';
+import '../../../components/Inputs/enum_field.dart';
+import '../../../components/Inputs/time_picker_rectangle.dart';
+import '../../../components/warning_modal.dart';
+import '../../AppConfig/change_password/change_password_screen.dart';
+import '../../AppConfig/change_phone/change_phone_screen.dart';
+import '../../AppConfig/deactivate_account/deactivate_account_screen.dart';
+import '../../AppConfig/delete_account/delete_warning_screen.dart';
+import '../../auth/initial_screen.dart';
 
 //TODO: gotta load the account configs
 class AppConfigScreen extends StatelessWidget {
